Updating fates admin to allow no gumballs for unreleased gumballs. Updated seeders to check if they should truncate.

+++ database/seeds/FactionsTableSeeder.php
@@ -15,7 +15,9 @@ class FactionsTableSeeder extends Seeder
      */
     public function run()
     {
-        $this->truncate();
+        if ((boolean) env('DB_TRUNCATE', false)) {
+            $this->truncate();
+        }
 
         $factions = [
             [
