+++ database/seeds/GroupsTableSeeder.php
@@ -15,7 +15,9 @@ class GroupsTableSeeder extends Seeder
      */
     public function run()
     {
-        $this->truncate();
+        if ((boolean) env('DB_TRUNCATE', false)) {
+            $this->truncate();
+        }
 
         $groups = [
             [
