+++ database/seeds/UsersTableSeeder.php
@@ -17,7 +17,9 @@ class UsersTableSeeder extends Seeder
      */
     public function run()
     {
-        $this->truncate();
+        if ((boolean) env('DB_TRUNCATE', false)) {
+            $this->truncate();
+        }
 
         $users = [
             [
